move onInstall code to browser.js
