+++ helper_js/browser.js
@@ -40,4 +40,60 @@ async function doConfirm(message) {
   });
 }
 
-export { doAlert, doConfirm }
+// On Install / upgrade / downgrade
+// Variable Initialization
+const install_message =
+  "Thank you for installing Sci-Hub X Now!\n" +
+  "Check out the options page to customize your experience! :)"
+const update_message =
+  "Thank you for upgrading Sci-Hub X Now!\n" +
+  "We have new features!\n" +
+  'Check out the "options" page now to enable them! :)';
+function onInstall(details) {
+  if (details.reason == "install") {
+    doAlert(install_message).then(() => { browser.runtime.openOptionsPage(); });
+    browser.tabs.create({ url: 'chrome://extensions/?options=' + chrome.runtime.id }).then();
+  }
+  if (details.reason == "update") {
+    function parseVersion(version) {
+      const parts = version.split(".");
+      const major = parseInt(parts[0], 10);
+      const minor = parseInt(parts[1], 10);
+      const bugfix = parseInt(parts[2], 10);
+
+      return { major, minor, bugfix };
+    }
+    let previousVersion = parseVersion(details.previousVersion);
+    let current_version = parseVersion(chrome.runtime.getManifest().version);
+
+    let major_update = current_version.major > previousVersion.major;
+    let minor_update =
+      current_version.major == previousVersion.major &&
+      current_version.minor > previousVersion.minor;
+    let bugfix_update =
+      current_version.major == previousVersion.major &&
+      current_version.minor == previousVersion.minor &&
+      current_version.bugfix > previousVersion.bugfix;
+    let no_update = current_version == previousVersion;
+
+    if (major_update) {
+      console.log("A major version update has occurred.");
+      doConfirm(update_message).then(() => { browser.runtime.openOptionsPage(); });
+    } else if (minor_update) {
+      console.log("A minor version update has occurred.");
+      doConfirm(update_message).then(() => { browser.runtime.openOptionsPage(); });
+    } else if (bugfix_update) {
+      console.log("A bugfix version update has occurred.");
+    } else if (no_update) {
+      console.log("No update has occurred.");
+    } else {
+      console.log("A downdate has occurred.");
+    }
+  }
+}
+
+function addListeners() {
+  chrome.runtime.onInstalled.addListener(onInstall);
+}
+
+export { doAlert, doConfirm, addListeners }
