+++ helper_js/config.js
@@ -28,15 +28,13 @@ function printVars() {
     "\nopenInNewTab: " + openInNewTab +
     "\nautoCheckServer: " + autoCheckServer);
 }
-function addListeners() {
-  chrome.storage.onChanged.addListener((changes, area) => {
-    if (area === 'local') {
-      for (let [key, { oldValue, newValue }] of Object.entries(changes)) {
-        setvariable(key, newValue);
-        console.log(`Changed "${key}" from "${oldValue}" to "${newValue}".`);
-      }
+function onChangedListener(changes, area) {
+  if (area === 'local') {
+    for (let [key, { oldValue, newValue }] of Object.entries(changes)) {
+      setvariable(key, newValue);
+      console.log(`Changed "${key}" from "${oldValue}" to "${newValue}".`);
     }
-  });
+  }
 }
 function setvariable(name, value) {
   switch (name) {
@@ -62,72 +60,31 @@ function setvariable(name, value) {
   console.log("setvariable called!!!");
   printVars();
 }
-function initialize(name, value) {
+function initializeVar(name, value) {
   console.log("initializing " + name + ": " + value);
   setvariable(name, value);
 }
 
-// Variable Initialization
-chrome.runtime.onInstalled.addListener(function (details) {
+chrome.storage.local.get(defaults, function (result) {
+  for (const property in result) {
+    console.log("Initializing ", property, ": ", result[property]);
+    initializeVar(property, result[property]);
+  }
+});
+
+
+// Initialization, on install / upgrade / downgrade
+function initialize(details) {
   // Set variables to default if they don't already exist
   chrome.storage.local.get(defaults, function (result) {
     console.log("Initializing variables onInstalled: ", result);
     chrome.storage.local.set(result); // for if variables were not set
   });
+}
 
-  if (details.reason == "install") {
-    browser.tabs.create({ url: 'chrome://extensions/?options=' + chrome.runtime.id }).then();
-  }
-  if (details.reason == "update") {
-    function parseVersion(version) {
-      const parts = version.split(".");
-      const major = parseInt(parts[0], 10);
-      const minor = parseInt(parts[1], 10);
-      const bugfix = parseInt(parts[2], 10);
-
-      return { major, minor, bugfix };
-    }
-    let previousVersion = parseVersion(details.previousVersion);
-    let current_version = parseVersion(chrome.runtime.getManifest().version);
-
-    let major_update = current_version.major > previousVersion.major;
-    let minor_update =
-      current_version.major == previousVersion.major &&
-      current_version.minor > previousVersion.minor;
-    let bugfix_update =
-      current_version.major == previousVersion.major &&
-      current_version.minor == previousVersion.minor &&
-      current_version.bugfix > previousVersion.bugfix;
-    let no_update = current_version == previousVersion;
-
-    let update_message =
-      "Thank you for upgrading Sci-Hub X Now!\n" +
-      "We have new features!\n" +
-      'Would you like to go to the "options" page now to enable them?';
-    if (major_update) {
-      console.log("A major version update has occurred.");
-      if (confirm(update_message)) {
-        browser.runtime.openOptionsPage();
-      }
-    } else if (minor_update) {
-      console.log("A minor version update has occurred.");
-      if (confirm(update_message)) {
-        browser.runtime.openOptionsPage();
-      }
-    } else if (bugfix_update) {
-      console.log("A bugfix version update has occurred.");
-    } else if (no_update) {
-      console.log("No update has occurred.");
-    } else {
-      console.log("A downdate has occurred.");
-    }
-  }
-});
-chrome.storage.local.get(defaults, function (result) {
-  for (const property in result) {
-    console.log("Initializing ", property, ": ", result[property]);
-    initialize(property, result[property]);
-  }
-});
+function addListeners() {
+  chrome.storage.onChanged.addListener(onChangedListener);
+  chrome.runtime.onInstalled.addListener(initialize);
+}
 
 export { doiRegex, sciHubUrl, autodownload, autoname, openInNewTab, autoCheckServer, venueAbbreviations, addListeners };
